Lazily initializing MPI to avoid cluttering the interface registration (#22)
In the old implementation, MPI was initialized when instantiating a
`mpi_communicator_backend` which in turn is called when populating an
`communicator_manager` in the plugin loading process. This may cause
significant lag during initialization. By lazily initializing MPI, we
intend to delay this lag to the moment where MPI is actually needed.

Unfortunately, the current criteria to determine whether MPI is
available (world size > 1) requires it to be initialized. Thus, this lag
is unavoidable whenever a communicator is needed, even if MPI is not
elected. In practice, when MPI is available, it will be the elected
communicator.